LayoutTests/platform/ios-simulator/editing/selection/select-line-break-with-opposite-directionality-expected.txt

This tests selecting LTR/RTL text that spans multiple lines in RTL/LTR block (opposite directionality). To manually test, select text in each of editable regions below by dragging mouse from one edge to another. The entire line should be selected.

TEST 1 line 1:FAIL - expected "abc" but got ""
TEST 1 line 2:FAIL - expected "def" but got ""
TEST 2 line 1:FAIL - expected "ابص" but got ""
TEST 3 line 1:FAIL - expected "abc" but got ""
TEST 4 line 1:FAIL - expected "ابص" but got ""
TEST 4 line 2:PASS
TEST 4 line 3:FAIL - expected "a" but got ""
TEST 5 line 1:FAIL - expected "abc" but got ""
TEST 5 line 2:PASS
TEST 5 line 3:FAIL - expected "ا" but got ""
TEST 6 line 1:FAIL - expected "ابص" but got ""
TEST 6 line 2:PASS
TEST 6 line 3:FAIL - expected "a" but got ""
TEST 7 line 1:FAIL - expected "ابص" but got ""
TEST 7 line 2:PASS
TEST 7 line 3:FAIL - expected "a" but got ""
TEST 8 line 1:FAIL - expected "ابص" but got ""
TEST 8 line 2:PASS
TEST 8 line 3:FAIL - expected "a" but got ""
TEST 9 line 1:FAIL - expected "abc" but got ""
TEST 9 line 2:PASS
TEST 9 line 3:FAIL - expected "a" but got ""
TEST 10 line 1:FAIL - expected "abc" but got ""
TEST 10 line 2:PASS
TEST 10 line 3:FAIL - expected "a" but got ""
